completed the download and registration of the files.
