+++ core_admin/coreAdmin/urls.py
@@ -49,7 +49,4 @@ urlpatterns += [
     path('admin/', admin.site.urls),
     url(r'^obtain-auth-token/$', csrf_exempt(obtain_auth_token)),
     url(r'^teste/', common_views.teste),
-    url(r'^teste2/', common_views.downloadOrbitalParameters),
-
-    url(r'^t1/', common_views.t1),
 ]
