+++ core_admin/orbit/bsp_jpl.py
@@ -12,8 +12,8 @@ from random import randrange
 from orbit.models import BspJplFile
 import logging
 from common.jsonfile import JsonFile
-
-class BSPJPL():
+from .download_parameters import DownloadParameters
+class BSPJPL(DownloadParameters):
     """
         Esta classe contem os metodos para baixar os arquivos BSP do JPL.
         O Download e feito usando um script disponibilizado pelo JPL,
@@ -190,57 +190,50 @@ class BSPJPL():
 
         JsonFile().write(steps, step_file)
 
-    def register_downloaded_files(self, records):
-        # Apos o Download e necessario o registro
-        self.logger.info("Register downloaded BSP JPL")
 
-        new = 0
-        updated = 0
-        downloaded = 0
-        not_downloaded = 0
-        for record in records:
-            if record.get("filename"):
-                downloaded += 1
+    def update_or_create_record(self, record):
+        return BspJplFile.objects.update_or_create(
+            name=record.get("name"),
+            defaults={
+                'filename': record.get("filename"),
+                'download_start_time': record.get("download_start_time"),
+                'download_finish_time': record.get("download_finish_time"),
+                'file_size': record.get("file_size"),
+            }
+        )
 
-                obj, created = BspJplFile.objects.update_or_create(
-                    name=record.get("name"),
-                    defaults={
-                        'filename': record.get("filename"),
-                        'download_start_time': record.get("download_start_time"),
-                        'download_finish_time': record.get("download_finish_time"),
-                        'file_size': record.get("file_size"),
-                    }
-                )
-
-                if created:
-                    new += 1
-                    self.logger.debug("Registered [ %s ] " % record.get("name"))
-                else:
-                    updated += updated
-                    self.logger.debug("Updated [ %s ] " % record.get("name"))
-            else:
-                not_downloaded += 1
-
-        self.logger.info("Inputs [ %s ] Downloaded [ %s ] Registered [ %s ] Updated [ %s ] Not Downloaded [ %s ]" % (
-        len(self.input_records), downloaded, new, updated, not_downloaded))
-
-    def read_input(self, input_file):
-
-        records = list()
-        with open(input_file) as csvfile:
-            reader = csv.DictReader(csvfile, delimiter=';')
-            for row in reader:
-
-                if row.get("need_download") in ['True', 'true', '1', 't', 'y', 'yes']:
-                    row["need_download"] = True
-
-                    self.need_download += 1
-
-                else:
-                    row["need_download"] = False
-
-                    self.not_need_download += 1
-
-                records.append(row)
-
-        return records
+    # def register_downloaded_files(self, records):
+    #     # Apos o Download e necessario o registro
+    #     self.logger.info("Register downloaded BSP JPL")
+    #
+    #     new = 0
+    #     updated = 0
+    #     downloaded = 0
+    #     not_downloaded = 0
+    #     for record in records:
+    #         if record.get("filename"):
+    #             downloaded += 1
+    #
+    #             obj, created = BspJplFile.objects.update_or_create(
+    #                 name=record.get("name"),
+    #                 defaults={
+    #                     'filename': record.get("filename"),
+    #                     'download_start_time': record.get("download_start_time"),
+    #                     'download_finish_time': record.get("download_finish_time"),
+    #                     'file_size': record.get("file_size"),
+    #                 }
+    #             )
+    #
+    #             if created:
+    #                 new += 1
+    #                 self.logger.debug("Registered [ %s ] " % record.get("name"))
+    #             else:
+    #                 updated += updated
+    #                 self.logger.debug("Updated [ %s ] " % record.get("name"))
+    #         else:
+    #             not_downloaded += 1
+    #
+    #     self.logger.info("Inputs [ %s ] Downloaded [ %s ] Registered [ %s ] Updated [ %s ] Not Downloaded [ %s ]" % (
+    #     len(self.input_records), downloaded, new, updated, not_downloaded))
+    #
+    #
